Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240120/합과 평균/sum-and-avg.cpp
@@ -7,7 +7,7 @@ int main() {
 
     cin >> a >> b;
 
-    cout << a+b << " " << (float)(a+b)/2;
+    cout << a+b << " " << (double)(a+b)/2;
 
     return 0;
 }
